Move frame's getbuffer functionality into a separate file
